Refactor deck list state model
