+++ src/features/deck-list/model/useDeckListState.ts
@@ -1,11 +1,10 @@
 import * as React from "react";
 
 import { useAuthUid } from "@/entities/auth";
-import { filterCardsByDeckId, type Card } from "@/entities/card";
+import { countCardsByDeckId, filterCardsByDeckId, type Card } from "@/entities/card";
 import { deleteDeck, mustFindDeckById, type Deck, type DeckId } from "@/entities/deck";
-import type { StudySession } from "@/entities/study-session";
+import { compareActiveDecks, groupDecksByStudyStatus, type StudySession } from "@/entities/study-session";
 import { downloadDeckCsv } from "../lib/deckCsv";
-import { buildDeckListSections } from "./buildDeckListSections";
 
 interface UseDeckListStateOptions {
   decks: Deck[];
@@ -18,8 +17,32 @@ interface DeletionTarget {
   cardCount: number;
 }
 
+export interface DeckListStudyProgress {
+  currentIndex: number;
+  cardCount: number;
+  lastStudiedAt: number;
+}
+
+export interface DeckListDeck {
+  id: DeckId;
+  name: string;
+  category: string;
+  isPublic: boolean;
+}
+
+export interface DeckListItem {
+  deck: DeckListDeck;
+  cardCount: number;
+  studyProgress?: DeckListStudyProgress;
+}
+
+export interface DeckListSections {
+  studying: DeckListItem[];
+  other: DeckListItem[];
+}
+
 export interface DeckListState {
-  sections: ReturnType<typeof buildDeckListSections>;
+  sections: DeckListSections;
   deletionTarget: { deckName: string; cardCount: number; hasError: boolean } | undefined;
   successMessage: string | undefined;
   onDownload: (id: DeckId) => void;
@@ -27,6 +50,44 @@ export interface DeckListState {
   onCancelDeletion: () => void;
   onConfirmDeletion: () => Promise<void>;
 }
+
+const compareDeckNames = (left: Deck, right: Deck): number => left.name.localeCompare(right.name);
+
+const createDeckListStudyProgress = (session: StudySession): DeckListStudyProgress => ({
+  currentIndex: session.currentIndex,
+  cardCount: session.cardOrderIds.length,
+  lastStudiedAt: session.lastStudiedAt,
+});
+
+const createDeckListDeck = (deck: Deck): DeckListDeck => ({
+  id: deck.id,
+  name: deck.name,
+  category: deck.category,
+  isPublic: deck.isPublic,
+});
+
+const buildDeckListSections = (
+  decks: Deck[],
+  cards: Card[],
+  sessionsByDeckId: Partial<Record<DeckId, StudySession>>
+): DeckListSections => {
+  const cardCounts = countCardsByDeckId(cards);
+  const createItem = (deck: Deck): DeckListItem => ({
+    deck: createDeckListDeck(deck),
+    cardCount: cardCounts.get(deck.id) ?? 0,
+  });
+  const { active: studyingDecks, inactive: otherDecks } = groupDecksByStudyStatus(decks, sessionsByDeckId);
+  studyingDecks.sort(compareActiveDecks);
+  otherDecks.sort(compareDeckNames);
+
+  return {
+    studying: studyingDecks.map(({ deck, session }) => ({
+      ...createItem(deck),
+      studyProgress: createDeckListStudyProgress(session),
+    })),
+    other: otherDecks.map(createItem),
+  };
+};
 
 export const useDeckListState = ({ decks, cards, sessionsByDeckId }: UseDeckListStateOptions): DeckListState => {
   const uid = useAuthUid();
